==> src/3_reduce_dimension_by_umap.py <==
import os

import numpy as np
import umap.umap_ as umap
from dotenv import load_dotenv

from codebase.dataprocessor.get_info import get_current_datetime
from codebase.dataprocessor.make_dir import make_dir

# 環境変数の読み込み
load_dotenv("/workspace/src/.env")
LOG_PATH: str = os.environ["LOG_PATH"]


def main():
    # load vectors
    vectors_path: str = f"{LOG_PATH}/embeddings/2023-6-10/13-6-29/vectors.csv"
    vectors: np.ndarray = np.loadtxt(vectors_path, delimiter=",")

    # reduce dimension
    n_epochs: int = 100
    n_neighbors: int = 15
    min_dist: float = 0.0
    metric: str = "cosine"
    verbose: bool = True
    model: umap.UMAP = umap.UMAP(
        n_epochs=n_epochs,
        n_neighbors=n_neighbors,
        min_dist=min_dist,
        metric=metric,
        verbose=verbose,
    )
    embedding: np.ndarray = model.fit_transform(vectors)

    # output results
    current_date, current_time = get_current_datetime()
    save_path: str = (
        f"{LOG_PATH}/reduce_dimension/{current_date}/{current_time}"
    )
    make_dir(save_path)
    np.savetxt(f"{save_path}/reduced_vectors.csv", embedding, delimiter=",")


if __name__ == "__main__":
    main()
